Resolve parser lang when different from filetype
Use nvim-treesitter API to resolve the parser language when it is
different than the buffer's `filetype`.

This happens for `tsx` files whose `filetype` is `typescriptreact` but
the Treesitter parser language is `tsx`. The mapping of filetypes and
parser languages is stored in nvim-treesitter [0].

PR #4 fixed that problem before. It seems that the solution stopped
working over time.

[0]: https://github.com/nvim-treesitter/nvim-treesitter/blob/32e364ea3c99aafcce2ce735fe091618f623d889/lua/nvim-treesitter/parsers.lua#L4-L21